+++ tests/reconnect_test.js
@@ -1,0 +1,27 @@
+var redis = require("redis").createClient();
+
+redis.on("error", function (err) {
+    console.log("Redis says: " + err);
+});
+
+redis.on("ready", function () {
+    console.log("Redis ready.");
+});
+
+redis.on("reconnecting", function (arg) {
+    console.log("Redis reconnecting: " + JSON.stringify(arg));
+});
+redis.on("connect", function () {
+    console.log("Redis connected.");
+});
+
+setInterval(function () {
+    var now = Date.now();
+    redis.set("now", now, function (err, res) {
+        if (err) {
+            console.log(now + " Redis reply error: " + err);
+        } else {
+            console.log(now + " Redis reply: " + res);
+        }
+    });
+}, 200);